move color out into color.h
