+++ romsel_dsimenutheme/arm9/source/graphics/ThemeTextures.h
@@ -302,66 +302,6 @@ private:
 
 //   xbbbbbgggggrrrrr according to http://problemkaputt.de/gbatek.htm#dsvideobgmodescontrol
 
-/**
- * Adapted from https://stackoverflow.com/questions/18937701/
- * applies alphablending with the given
- * RGB555 foreground, RGB555 background, and alpha from
- * 0 to 128 (0, 1.0).
- * The lower the alpha the more transparent, but
- * this function does not produce good results at the extremes
- * (near 0 or 128).
- */
-inline u16 alphablend(const u16 fg, const u16 bg, const u8 alpha) {
-
-	  u8 fg_b, fg_g, fg_r, bg_b, bg_g, bg_r;
-		fg_b = ((fg) >> 10) & 31;
-		fg_g = ((fg) >> 5) & 31;
-		fg_r = (fg)&31;
-    
-		bg_b = ((fg) >> 10) & 31;
-		bg_g = ((fg) >> 5) & 31;
-		bg_r = (fg)&31;
-
-
- // Alpha blend components
-    unsigned out_r = fg_r * alpha + bg_r * (255 - alpha);
-    unsigned out_g = fg_g * alpha + bg_g * (255 - alpha);
-    unsigned out_b = fg_b * alpha + bg_b * (255 - alpha);
-    out_r = (out_r + 1 + (out_r >> 8)) >> 8;
-    out_g = (out_g + 1 + (out_g >> 8)) >> 8;
-    out_b = (out_b + 1 + (out_b >> 8)) >> 8;
-    return (u16) ((out_b << 10) | (out_g << 5) | out_r | BIT(15));
-}
-
-
-
-/**
- * Adapted from https://stackoverflow.com/questions/18937701/
- * applies alphablending with the given
- * RGB555 foreground, RGB555 background, and alpha from
- * 0 to 128 (0, 1.0).
- * The lower the alpha the more transparent, but
- * this function does not produce good results at the extremes
- * (near 0 or 128).
- */
-inline u16 blend(const u16 fg, const u16 bg) {
-
-	  u8 fg_b, fg_g, fg_r, bg_b, bg_g, bg_r;
-		fg_b = ((fg) >> 10) & 31;
-		fg_g = ((fg) >> 5) & 31;
-		fg_r = (fg)&31;
-    
-		bg_b = ((fg) >> 10) & 31;
-		bg_g = ((fg) >> 5) & 31;
-		bg_r = (fg)&31;
-
-
- // Alpha blend components
-    unsigned out_r = min((fg_r + bg_r) >> 1, 255);
-    unsigned out_g = min((fg_g + bg_g) >> 1, 255);
-    unsigned out_b = min((fg_b + bg_b) >> 1, 255);
-    return (u16) ((out_b << 10) | (out_g << 5) | out_r | BIT(15));
-}
 
 typedef singleton<ThemeTextures> themeTextures_s;
 inline ThemeTextures &tex() { return themeTextures_s::instance(); }
